feat: refactored skip, take and response
