+++ Dockerfile
@@ -1,16 +1,14 @@
 FROM node:22.4.0-alpine3.20 AS base
 
 #This is to use pnpm from your local machine
-#ENV PNPM_HOME="/pnpm"
-#ENV PATH="$PNPM_HOME:$PATH"
-RUN npm install -g pnpm && pnpm --version
+ENV PNPM_HOME="/pnpm"
+ENV PATH="$PNPM_HOME:$PATH"
 #corepack is for enabling symlink as it is experimental
 RUN corepack enable
 
 WORKDIR /app
 
-COPY package*.json ./
-COPY pnpm-lock.yaml ./
+COPY package*.json pnpm-lock.yaml ./
 RUN pnpm install || cat /root/.pnpm-debug.log
 
 COPY . .
