+++ app/api/swagger/route.js
@@ -1,4 +1,4 @@
-import swaggerSpec from "@/app/lib/swagger";
+import swaggerSpec from "@/lib/swagger";
 
 export async function GET(request) {
   return new Response(JSON.stringify(swaggerSpec), {
