+++ lib/swagger.js
