+++ utils/API/skip.js
@@ -1,0 +1,4 @@
+export const skip = (searchParams) =>
+  parseInt(searchParams.get("skip"), 10)
+    ? parseInt(searchParams.get("skip"), 10)
+    : 0;
