+++ utils/API/take.js
@@ -1,0 +1,4 @@
+export const take = (searchParams) =>
+  parseInt(searchParams.get("take"), 10)
+    ? parseInt(searchParams.get("take"), 10)
+    : 10;
